chore: restructure components
