+++ src/App.tsx
@@ -2,13 +2,13 @@ import { useEffect, useState, useRef, useCallback } from "react";
 import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
 import Lenis from "@studio-freight/lenis";
-import { Header } from "./components/Header/Header";
+import { Header } from "./components/layout/Header/Header";
 import { HeroSection } from "./components/HeroSection/HeroSection";
 import { RoadmapSection } from "./components/RoadmapSection/RoadmapSection";
 import { AboutSection } from "./components/AboutSection/AboutSection";
 import { CommunitySection } from "./components/CommunitySection/CommunitySection";
 import { FAQSection } from "./components/FAQSection/FAQSection";
-import { Footer } from "./components/Footer/Footer";
+import { Footer } from "./components/layout/Footer/Footer";
 import { ThreeDBackground } from "./components/ThreeDBackground/ThreeDBackground";
 import "./styles/globals.scss";
 import usePrefersReducedMotion from "./hooks/usePrefersReducedMotion";
