+++ src/app/(marketing)/page.tsx
@@ -1,0 +1,3 @@
+export default function MarketingPage() {
+  return <div></div>;
+}
