+++ src/components/primitives/Card.module.scss
@@ -1,0 +1,6 @@
+.card {
+  background-color: var(--color-surface-secondary);
+  border-radius: var(--border-radius-lg);
+  padding: var(--spacing-1-5);
+  box-shadow: var(--shadow-lg);
+}
