+++ src/components/primitives/Input.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import styles from './Input.module.scss';
+
+export const Input: React.FC<React.InputHTMLAttributes<HTMLInputElement>> = ({
+  className = '',
+  ...props
+}) => {
+  return <input className={`${styles.input} ${className}`} {...props} />;
+};
